requirements.txt
SimpleITK>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
tqdm>=4.66.0
PyYAML>=6.0
pydicom>=2.4.0
pytest>=7.4.0
